+++ gantry/common/util.py
@@ -1,5 +1,8 @@
+import platform
+import tempfile
 import time
-from typing import TYPE_CHECKING, Iterable, Tuple
+from pathlib import Path
+from typing import TYPE_CHECKING, Iterable, List, Optional, Tuple
 
 import rich
 from beaker import (
@@ -8,15 +11,18 @@ from beaker import (
     DatasetConflict,
     DatasetNotFound,
     Digest,
+    ExperimentSpec,
     SecretNotFound,
     TaskResources,
+    TaskSpec,
 )
 from rich import print
 from rich.console import Console
 
 from ..exceptions import *
 from ..version import VERSION
-from .constants import ENTRYPOINT, GITHUB_TOKEN_SECRET
+from . import constants
+from .aliases import PathOrStr
 
 if TYPE_CHECKING:
     from datetime import timedelta
@@ -101,51 +107,66 @@ def ensure_repo(allow_dirty: bool = False) -> Tuple[str, str, str]:
 
 def ensure_entrypoint_dataset(beaker: Beaker) -> Dataset:
     import hashlib
+    from importlib.resources import read_binary
 
-    from importlib_resources import as_file, files
+    import gantry
 
     workspace_id = beaker.workspace.get().id
 
-    with as_file(files("gantry").joinpath(ENTRYPOINT)) as entrypoint_path:
-        # Get hash of the local entrypoint source file.
-        sha256_hash = hashlib.sha256()
-        with open(entrypoint_path, "rb") as f:
-            sha256_hash.update(f.read())
+    # Get hash of the local entrypoint source file.
+    sha256_hash = hashlib.sha256()
+    contents = read_binary(gantry, constants.ENTRYPOINT)
+    tag_start = contents.find(b"${{")
+    while tag_start != -1:
+        tag_end = contents.find(b"}}") + 2
+        tag = contents[tag_start:tag_end]
+        constant_name = tag.split(b" ")[1].decode()
+        contents = contents.replace(tag, getattr(constants, constant_name).encode())  # type: ignore
+        tag_start = contents.find(b"${{", tag_end)
+    assert b"${{" not in contents
+    sha256_hash.update(contents)
 
-        entrypoint_dataset_name = f"gantry-v{VERSION}-{workspace_id}-{sha256_hash.hexdigest()[:6]}"
+    entrypoint_dataset_name = f"gantry-v{VERSION}-{workspace_id}-{sha256_hash.hexdigest()[:6]}"
 
-        # Ensure gantry entrypoint dataset exists.
-        gantry_entrypoint_dataset: Dataset
+    # Ensure gantry entrypoint dataset exists.
+    gantry_entrypoint_dataset: Dataset
+    try:
+        gantry_entrypoint_dataset = beaker.dataset.get(entrypoint_dataset_name)
+    except DatasetNotFound:
+        # Create it.
+        print(f"Creating entrypoint dataset '{entrypoint_dataset_name}'")
         try:
-            gantry_entrypoint_dataset = beaker.dataset.get(entrypoint_dataset_name)
-        except DatasetNotFound:
-            # Create it.
-            print(f"Creating entrypoint dataset '{entrypoint_dataset_name}'")
-            try:
+            with tempfile.TemporaryDirectory() as tmpdirname:
+                tmpdir = Path(tmpdirname)
+                entrypoint_path = tmpdir / constants.ENTRYPOINT
+                with open(entrypoint_path, "wb") as entrypoint_file:
+                    entrypoint_file.write(contents)
                 gantry_entrypoint_dataset = beaker.dataset.create(
                     entrypoint_dataset_name, entrypoint_path
                 )
-            except DatasetConflict:  # could be in a race with another `gantry` process.
-                time.sleep(1.0)
-                gantry_entrypoint_dataset = beaker.dataset.get(entrypoint_dataset_name)
+        except DatasetConflict:  # could be in a race with another `gantry` process.
+            time.sleep(1.0)
+            gantry_entrypoint_dataset = beaker.dataset.get(entrypoint_dataset_name)
 
-        # Verify contents.
-        err_msg = (
-            f"Checksum failed for entrypoint dataset {beaker.dataset.url(gantry_entrypoint_dataset)}\n"
-            f"This could be a bug, or it could mean someone has tampered with the dataset.\n"
-            f"If you're sure no one has tampered with it, you can delete the dataset from "
-            f"the Beaker dashboard and try again."
-        )
-        ds_files = list(beaker.dataset.ls(gantry_entrypoint_dataset))
-        if len(ds_files) != 1:
-            raise EntrypointChecksumError(err_msg)
-        if ds_files[0].digest != Digest(sha256_hash.digest()):
-            raise EntrypointChecksumError(err_msg)
+    # Verify contents.
+    err_msg = (
+        f"Checksum failed for entrypoint dataset {beaker.dataset.url(gantry_entrypoint_dataset)}\n"
+        f"This could be a bug, or it could mean someone has tampered with the dataset.\n"
+        f"If you're sure no one has tampered with it, you can delete the dataset from "
+        f"the Beaker dashboard and try again."
+    )
+    ds_files = list(beaker.dataset.ls(gantry_entrypoint_dataset))
+    if len(ds_files) != 1:
+        raise EntrypointChecksumError(err_msg)
+    if ds_files[0].digest != Digest(sha256_hash.digest()):
+        raise EntrypointChecksumError(err_msg)
 
-        return gantry_entrypoint_dataset
+    return gantry_entrypoint_dataset
 
 
-def ensure_github_token_secret(beaker: Beaker, secret_name: str = GITHUB_TOKEN_SECRET) -> str:
+def ensure_github_token_secret(
+    beaker: Beaker, secret_name: str = constants.GITHUB_TOKEN_SECRET
+) -> str:
     try:
         beaker.secret.get(secret_name)
     except SecretNotFound:
@@ -200,3 +221,62 @@ def format_timedelta(td: "timedelta") -> str:
     if seconds:
         parts.append(format_value_and_unit(seconds, "second"))
     return ", ".join(parts)
+
+
+def build_experiment_spec(
+    task_name: str,
+    cluster_to_use: str,
+    task_resources: TaskResources,
+    arguments: List[str],
+    entrypoint_dataset: str,
+    github_account: str,
+    github_repo: str,
+    git_ref: str,
+    description: Optional[str] = None,
+    beaker_image: Optional[str] = None,
+    docker_image: Optional[str] = None,
+    gh_token_secret: Optional[str] = constants.GITHUB_TOKEN_SECRET,
+    conda: Optional[PathOrStr] = None,
+    pip: Optional[PathOrStr] = None,
+    nfs: Optional[bool] = None,
+):
+    task_spec = (
+        TaskSpec.new(
+            task_name,
+            cluster_to_use,
+            beaker_image=beaker_image,
+            docker_image=docker_image,
+            result_path=constants.RESULTS_DIR,
+            command=["bash", "/gantry/entrypoint.sh"],
+            arguments=arguments,
+            resources=task_resources,
+        )
+        .with_env_var(name="GANTRY_VERSION", value=VERSION)
+        .with_env_var(name="GITHUB_TOKEN", secret=gh_token_secret)
+        .with_env_var(name="GITHUB_REPO", value=f"{github_account}/{github_repo}")
+        .with_env_var(name="GIT_REF", value=git_ref)
+        .with_dataset("/gantry", beaker=entrypoint_dataset)
+    )
+
+    if conda is not None:
+        task_spec = task_spec.with_env_var(
+            name="CONDA_ENV_FILE",
+            value=str(conda),
+        )
+    elif not Path(constants.CONDA_ENV_FILE).is_file():
+        task_spec = task_spec.with_env_var(name="PYTHON_VERSION", value=platform.python_version())
+
+    if pip is not None:
+        task_spec = task_spec.with_env_var(
+            name="PIP_REQUIREMENTS_FILE",
+            value=str(pip),
+        )
+
+    if nfs is None:
+        if cluster_to_use in constants.NFS_SUPPORTED_CLUSTERS:
+            nfs = True
+
+    if nfs:
+        task_spec = task_spec.with_dataset(constants.NFS_MOUNT, host_path=constants.NFS_MOUNT)
+
+    return ExperimentSpec(description=description, tasks=[task_spec])